+++ tests/PdfDocOutputTest.php
@@ -15,16 +15,18 @@ namespace fpdf\Tests;
 
 use fpdf\Enums\PdfDestination;
 use fpdf\PdfException;
+use fpdf\Tests\Fixture\PdfSapiHeadersDocument;
 
 class PdfDocOutputTest extends AbstractPdfDocTestCase
 {
     public function testOutputDownloadInvalid(): void
     {
+        $expected = 'fake';
         self::expectException(PdfException::class);
         self::expectExceptionMessage('Some data has already been output, can not send PDF file.');
-        $this->expectOutputString('fake');
+        $this->expectOutputString($expected);
         $doc = $this->createDocument();
-        echo 'fake';
+        echo $expected;
         $doc->output(PdfDestination::DOWNLOAD);
     }
 
@@ -41,7 +43,7 @@ class PdfDocOutputTest extends AbstractPdfDocTestCase
         self::expectException(PdfException::class);
         self::expectExceptionMessageMatches('/Unable to create output file:.*/');
         $doc = $this->createDocument();
-        $doc->output(PdfDestination::FILE, 'https://www.bibi.nu/file.doc');
+        $doc->output(PdfDestination::FILE, '///doc.pdf');
     }
 
     public function testOutputInlineInvalid(): void
@@ -67,5 +69,27 @@ class PdfDocOutputTest extends AbstractPdfDocTestCase
         $doc = $this->createDocument();
         $value = $doc->output(PdfDestination::STRING);
         self::assertNotEmpty($value);
+    }
+    public function testOutputWithBomContent(): void
+    {
+        echo "\xEF\xBB\xBF";
+        $doc = $this->createDocument();
+        $doc->output(PdfDestination::DOWNLOAD);
+        self::assertSame(1, $doc->getPage());
+    }
+    public function testOutputWithHeadersSent(): void
+    {
+        $doc = new PdfSapiHeadersDocument(false, true);
+        self::expectException(PdfException::class);
+        self::expectExceptionMessage('Some data has already been output, can not send PDF file (output started at code.php:10).');
+        $doc->output(PdfDestination::DOWNLOAD);
+    }
+
+    public function testOutputWithSpaceContent(): void
+    {
+        echo "\r\n\t\f\v ";
+        $doc = $this->createDocument();
+        $doc->output(PdfDestination::DOWNLOAD);
+        self::assertSame(1, $doc->getPage());
     }
 }